requirements.txt
pillow
python-dotenv
google-generativeai
